IndexTweetLambda/requirements.txt
boto3==1.4.8
botocore==1.8.1
certifi==2017.11.5
chardet==3.0.4
docutils==0.14
elasticsearch==6.0.0
idna==2.6
jmespath==0.9.3
kafka-python==1.3.5
python-dateutil==2.6.1
requests==2.18.4
s3transfer==0.1.11
six==1.11.0
urllib3==1.22
